- Refactored components to Individual files.
