+++ components/mdx/components/quiz/slot.tsx
@@ -1,0 +1,5 @@
+import { ComponentWithChildren } from './utils';
+
+export function Slot({ children }: ComponentWithChildren) {
+    return <>{children}</>;
+}
