+++ components/mdx/components/quiz/utils.ts
@@ -1,0 +1,3 @@
+export interface ComponentWithChildren {
+    children: React.ReactNode;
+}
