+++ pages/quiz/[slug].tsx
@@ -24,7 +24,6 @@ export default function Quiz({ quiz }: QuizProps) {
 
     return (
         <Layout>
-            {/* <Example /> */}
             <QuizHeader title={title} />
             <QuizBody content={content} />
         </Layout>
